requirements.txt
pygame==2.1.2
tensorflow==2.6.2
keras==2.6.0
twilio==7.9.2
scikit-learn==1.0.2
Pillow==9.1.0
pandas==1.1.5
numpy==1.19.5